Client-side validation of email removed
The regular expression used for that purpose was only able to see if
the email field was blank or not, and the validation caused problems
with signup forms added to the theme. There's already server-side
validation for that. Also, I added styles to form fields in error and
improved error feedback.