Added Support for Line Item Properties
Take note that although I added support also for file uploads, file
uploads won't work unless Ajax is disabled. In the Knowledge Base FAQ
on the Radiance theme, I show how to disable Ajax in the Radiance theme
so I am leaving that code in. It does not hurt anything.